Extract ProductDetail styles into a new JS file.
